javascriptを/static/js/buttonChange.jsに移動
