+++ meiboWeb/templates/default-layout.hamlet
@@ -22,6 +22,7 @@
 
 <!-- Page Contents -->
 
+<!--
 <div .container>
     $if not $ Just HomeR == mcurrentRoute
         <ul .breadcrumb>
@@ -33,6 +34,7 @@
 
     $maybe msg <- mmsg
       <div .alert.alert-info #message>#{msg}
+-->
 
 ^{widget}
